Update patch set 1
Patch Set 1: Code-Review-1

WIP. Not all devices that use shield-common need this, so some form of filtering is needed.

This code rotates touch based on the nvidia set top box property. When in stb mode, shieldtablet turns the built in display off, but forces it to 0 degree rotation. However, the external display is in landscape, so touch really should be forced to 90 degree rotation. This is the stock behavior.

Patch-set: 1
Label: Code-Review=-1